+++ src/Events/SubscriptionCreated.php
@@ -9,6 +9,7 @@ use OfficeGuy\LaravelSumitGateway\Models\Subscription;
 class SubscriptionCreated
 {
     public function __construct(
-        public Subscription $subscription
+        public Subscription $subscription,
+        public array $response = []
     ) {}
 }